backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
boto3==1.29.7
celery==5.3.4
redis==5.0.1
numpy==1.24.3
torch==2.1.0
torchvision==0.16.0
librosa==0.10.1
opencv-python-headless==4.8.1.78
openai-whisper==20231117
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

